RavenDB-18645 - fixes for Get /collections/docs
