+++ src/Raven.Server/Documents/Sharding/Handlers/Processors/Collections/ShardedCollectionsHandlerProcessorForGetCollectionDocuments.cs
@@ -1,17 +1,20 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Threading;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.Http;
 using Raven.Client.Documents.Commands;
 using Raven.Client.Documents.Session.Operations;
+using Raven.Client.Exceptions.Sharding;
 using Raven.Client.Http;
 using Raven.Server.Documents.Handlers.Processors.Collections;
 using Raven.Server.Documents.Sharding.Handlers.ContinuationTokens;
 using Raven.Server.Documents.Sharding.Operations;
 using Raven.Server.Documents.Sharding.Streaming;
-using Raven.Server.NotificationCenter.Notifications.Details;
+using Raven.Server.Json;
 using Raven.Server.ServerWide.Context;
+using Sparrow.Json;
 using Sparrow.Utils;
 
 namespace Raven.Server.Documents.Sharding.Handlers.Processors.Collections
@@ -22,31 +25,50 @@ namespace Raven.Server.Documents.Sharding.Handlers.Processors.Collections
         {
         }
 
-        protected override async ValueTask<IAsyncEnumerable<Document>> GetCollectionDocumentsAsync(TransactionOperationContext context, string name, int start, int pageSize)
+        protected override async ValueTask GetCollectionDocumentsAndWriteAsync(TransactionOperationContext context, string name, int _, int __, CancellationToken token)
         {
-            var token = RequestHandler.ContinuationTokens.GetOrCreateContinuationToken(context);
+            throw new NotSupportedInShardingException($"Fetching documents by collection not supported in sharding.");
+
+            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Karmel, DevelopmentHelper.Severity.Normal,
+                "Need to have ShardedStreamDocumentsOperation fetch by collection name");
 
             DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Arek, DevelopmentHelper.Severity.Normal,
                 "See `null` passed as etag to above new ShardedCollectionPreviewOperation()");
 
-            var op = new ShardedStreamDocumentsOperation(HttpContext, null, token);
+            var continuationToken = RequestHandler.ContinuationTokens.GetOrCreateContinuationToken(context);
+
+            var op = new ShardedStreamDocumentsOperation(HttpContext, null, continuationToken);
             var results = await RequestHandler.ShardExecutor.ExecuteParallelForAllAsync(op);
             using var streams = await results.Result.InitializeAsync(RequestHandler.DatabaseContext, HttpContext.RequestAborted);
-            return GetDocuments(streams, token);
-        }
+            var documents = GetDocuments(streams, continuationToken);
 
+            long numberOfResults;
+            long totalDocumentsSizeInBytes;
+
+            await using (var writer = new AsyncBlittableJsonTextWriter(context, RequestHandler.ResponseBodyStream()))
+            {
+                writer.WriteStartObject();
+                writer.WritePropertyName("Results");
+                (numberOfResults, totalDocumentsSizeInBytes) = await writer.WriteDocumentsAsync(context, documents, metadataOnly: false, token);
+
+                if (continuationToken != null)
+                {
+                    writer.WriteComma();
+                    writer.WriteContinuationToken(context, continuationToken);
+                }
+                writer.WriteEndObject();
+            }
+
+            //AddPagingPerformanceHint(PagingOperationType.Documents, "Collection", HttpContext.Request.QueryString.Value, numberOfResults, pageSize, sw.ElapsedMilliseconds, totalDocumentsSizeInBytes);
+            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Karmel, DevelopmentHelper.Severity.Normal, "Handle sharded AddPagingPerformanceHint");
+        }
+        
         public async IAsyncEnumerable<Document> GetDocuments(CombinedReadContinuationState documents, ShardedPagingContinuation pagingContinuation)
         {
             await foreach (var result in RequestHandler.DatabaseContext.Streaming.PagedShardedDocumentsByLastModified(documents, nameof(CollectionResult.Results), pagingContinuation))
             {
                 yield return result.Item;
             }
-        }
-
-        protected override void AddPagingPerformanceHint(PagingOperationType operation, string action, string details, long numberOfResults, int pageSize, long duration,
-            long totalDocumentsSizeInBytes)
-        {
-            DevelopmentHelper.ShardingToDo(DevelopmentHelper.TeamMember.Karmel, DevelopmentHelper.Severity.Normal, "Handle sharded AddPagingPerformanceHint");
         }
     }
